Move src into own folder (#295)
