+++ example/package.json
@@ -24,7 +24,7 @@
     "env-cmd": "^5.0.0",
     "imports-loader": "^0.7.1",
     "jest": "^21.2.1",
-    "serverless-appsync-plugin": "file:./..",
+    "serverless-appsync-plugin": "file:./../src",
     "serverless-webpack": "^4.1.0",
     "webpack": "^3.8.1",
     "webpack-node-externals": "^1.6.0"
