+++ package.json
@@ -2,9 +2,12 @@
   "name": "serverless-appsync-plugin",
   "version": "1.1.2",
   "description": "AWS AppSync support for the Serverless Framework",
-  "main": "index.js",
+  "main": "src/index.js",
   "repository": "https://github.com/sid88in/serverless-appsync-plugin",
   "license": "MIT",
+  "files": [
+    "/src"
+  ],
   "scripts": {
     "test": "jest",
     "test:watch": "jest --watch",
