+++ src/get-config.js
