+++ src/get-stack-value.js
